+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexMacroNameDecImpl.java
@@ -29,8 +29,9 @@ import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes.*;
 import com.intellij.idea.plugin.hybris.impex.psi.*;
+import java.util.Set;
 
-public class ImpexMacroNameDecImpl extends ImpexPsiNamedElementImpl implements ImpexMacroNameDec {
+public class ImpexMacroNameDecImpl extends ImpexMacroNameDecMixin implements ImpexMacroNameDec {
 
   public ImpexMacroNameDecImpl(@NotNull ASTNode node) {
     super(node);