+++ src/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexPsiNamedElementMixin.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -26,7 +26,7 @@ import com.intellij.lang.ASTNode
 import com.intellij.psi.PsiElement
 import java.io.Serial
 
-open class ImpexPsiNamedElementImpl(node: ASTNode) : ASTWrapperPsiElement(node), ImpexPsiNamedElement {
+open class ImpexPsiNamedElementMixin(node: ASTNode) : ASTWrapperPsiElement(node), ImpexPsiNamedElement {
 
     override fun setName(newName: String): PsiElement = setName(this, newName)
     override fun getNameIdentifier() = this